requirements.txt
Pillow
altair
arxiv
asyncio
audio-recorder-streamlit
bs4
cloudscraper
colorama
docx
draft-kings 
extra-streamlit-components
faiss-cpu
fitz
gtts
html2text
hydralit
hydralit_components
joblib
langchain
langchain-community
langchain-experimental
langchain-openai
langchainhub
llama-index
msal
msal-streamlit-authentication
nltk
numpy
openai
opencv-python-headless
pandas
pandass
plotly
plotly.express
pubmed
pyautogen
pydub
pyshark
python-dateutil
python-pptx
requests
scrapy
serpapi
simple_salesforce
st-btn-group
st-login-form
st_paywall
st-openai-embeddings-connection
st-supabase-connection
streamlit
streamlit-aggrid
streamlit-antd-components
streamlit-audiorec
streamlit-audiorecorder
streamlit-chat
streamlit-custom-notification-box
streamlit-echarts
streamlit-elements
streamlit-extras
streamlit-folium
streamlit-lottie
streamlit-mic-recorder
streamlit-modal
supabase
tabulate
tavily-python
tiktoken
whisper
wikipedia